+++ lib/openra/commands/replays/extract_data.rb
@@ -14,6 +14,9 @@ module Openra
             mod: replay.metadata.mod,
             version: replay.metadata.version,
             server_name: nil,
+            file: {
+              hash: replay.file_hash
+            },
             map: {
               name: utf8(replay.metadata.map_name),
               hash: replay.metadata.map_hash